Update README / Reusable animations collection
I’ve shared react-animations as a component collection with Bit, which is a tool for publishing multiple components from any project to be installed and even developed from any other project. This opens a few new possibilities for people to engage with the animations:

1. Explore the animations with visual previews in a gallery collection
2. Install each animation with npm/yarn as an independent package
3. Use bit to develop each component right from their own project
4. Play with each animation in a live playground

I think it’s a valuable addition to this great library. I’ll be very happy to hand over ownership for the collection and help maintain it if you like.  Cheers